+++ config/routes.rb
@@ -47,6 +47,9 @@ Rails.application.routes.draw do
   resources :organizations, only: %i[index new create show edit update destroy] do
     member { post :add_member }
     member { post :undelete }
+    member { post :import_students }
+    member { post :confirm_import }
+    member { get :import }
   end
   resources :chapters, only: %i[index new create show edit update destroy] do
     member { post :undelete }
@@ -54,6 +57,8 @@ Rails.application.routes.draw do
 
   resources :groups, only: %i[index new create show edit update destroy] do
     member { post :undelete }
+    member { get :enroll_students }
+    member { post :confirm_enrollments }
 
     resources :students, controller: :group_students, only: %i[new create edit update] do
       member { post :cancel_edit }
@@ -62,6 +67,7 @@ Rails.application.routes.draw do
 
   resources :students, only: %i[index new create show edit update destroy] do
     member { post :undelete }
+
     collection { get '/mlid/:organization_id', to: 'students#mlid' }
 
     resources :student_images, only: %i[create destroy]