+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 - Added dropdown to select enrollment starting date for new students in a group
+- Added ability to enroll students with no active enrollments in a group
+- Added ability to import students to an organization from `.csv` files
 - Added organization statistics in separate component when viewing an organization
 
 ## 0.37.0 - Student Enrollments Overhaul